remove redundant lines
