+++ vllm_ascend/worker/npu_input_batch.py
@@ -26,8 +26,17 @@ from vllm.v1.sample.logits_processor import (BatchUpdateBuilder,
                                              LogitsProcessors)
 from vllm.v1.worker.gpu_input_batch import InputBatch
 
-from vllm_ascend.pool.medatata import PoolingStates
 from vllm_ascend.worker.block_table import MultiGroupBlockTable
+
+
+class PoolingStates:
+    # NOTE: This should be removed after we drop support of vLLM v0.12.0
+    def __init__(self):
+        # for chunked prefill with ALL pooling
+        self.hidden_states_cache: list[torch.Tensor] = []
+
+    def clean(self):
+        self.hidden_states_cache.clear()
 
 
 class NPUInputBatch(InputBatch):
